--- a/pkg/tsdb/influxdb/influxql/util/util.go
+++ b/pkg/tsdb/influxdb/influxql/util/util.go
@@ -98,23 +98,6 @@ func BuildFrameNameFromQuery(rowName, column string, tags map[string]string, fra
 }
 
 func ParseTimestamp(value any) (time.Time, error) {
-	timestampNumber, ok := value.(json.Number)
-	if !ok {
-		return time.Time{}, fmt.Errorf("timestamp-value has invalid type: %#v", value)
-	}
-	timestampInMilliseconds, err := timestampNumber.Int64()
-	if err != nil {
-		return time.Time{}, err
-	}
-
-	// currently in the code the influxdb-timestamps are requested with
-	// milliseconds-precision, meaning these values are milliseconds
-	t := time.UnixMilli(timestampInMilliseconds).UTC()
-
-	return t, nil
-}
-
-func ParseTimestamp2(value any) (time.Time, error) {
 	timestampNumber, ok := value.(float64)
 	if !ok {
 		return time.Time{}, fmt.Errorf("timestamp-value has invalid type: %#v", value)