**What is this feature?**

Small response (9 values each value has 17 items ) difference

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/grafana/pkg/tsdb/influxdb/influxql
             │ buffered.txt │            buffered3.txt            │
             │    sec/op    │   sec/op     vs base                │
ParseJson-10   119.97µ ± 0%   91.65µ ± 1%  -23.61% (p=0.000 n=10)

             │ buffered.txt │            buffered3.txt            │
             │     B/op     │     B/op      vs base               │
ParseJson-10   123.7Ki ± 0%   117.3Ki ± 0%  -5.17% (p=0.000 n=10)

             │ buffered.txt │           buffered3.txt            │
             │  allocs/op   │  allocs/op   vs base               │
ParseJson-10    2.527k ± 0%   2.510k ± 0%  -0.67% (p=0.000 n=10)

```

Big response (55318 values each value has 2 items) difference:

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/grafana/pkg/tsdb/influxdb/influxql
             │ buffered.txt │            buffered3.txt            │
             │    sec/op    │   sec/op     vs base                │
ParseJson-10    23.25m ± 0%   10.96m ± 0%  -52.86% (p=0.000 n=10)

             │ buffered.txt │            buffered3.txt             │
             │     B/op     │     B/op      vs base                │
ParseJson-10   26.62Mi ± 0%   21.23Mi ± 0%  -20.24% (p=0.000 n=10)

             │ buffered.txt │            buffered3.txt            │
             │  allocs/op   │  allocs/op   vs base                │
ParseJson-10    386.9k ± 0%   278.0k ± 0%  -28.13% (p=0.000 n=10)

```